Add: Display cards for the list of orders made by the user.
Firstly, the useEffect hook in ordersList was not displaying the page since, the id when the user was not logged in,
added an comparison statement to use the hook only in the case where the user was logged in. Then, added
Cards from material-ui gave them equal spacing. Added a button which can later be modified to display the values in the map itself.